Made groupAdmin boolean it's own component that is accessible globally
